+++ tokio-util/src/sync/mod.rs
@@ -3,8 +3,6 @@
 mod cancellation_token;
 pub use cancellation_token::{guard::DropGuard, CancellationToken, WaitForCancellationFuture};
 
-mod intrusive_double_linked_list;
-
 mod mpsc;
 pub use mpsc::{PollSendError, PollSender};