+++ tokio/src/runtime/mod.rs
@@ -216,6 +216,9 @@ cfg_rt! {
 
     mod builder;
     pub use self::builder::Builder;
+    cfg_unstable! {
+        pub use self::builder::UnhandledPanic;
+    }
 
     pub(crate) mod context;
     mod driver;
@@ -435,6 +438,8 @@ cfg_rt! {
         /// When the multi thread scheduler is used this will allow futures
         /// to run within the io driver and timer context of the overall runtime.
         ///
+        /// Any spawned tasks will continue running after `block_on` returns.
+        ///
         /// # Current thread scheduler
         ///
         /// When the current thread scheduler is enabled `block_on`
@@ -443,6 +448,9 @@ cfg_rt! {
         /// other threads which do not own the drivers will hook into that one.
         /// When the first `block_on` completes, other threads will be able to
         /// "steal" the driver to allow continued execution of their futures.
+        ///
+        /// Any spawned tasks will be suspended after `block_on` returns. Calling
+        /// `block_on` again will resume previously spawned tasks.
         ///
         /// # Panics
         ///
@@ -464,7 +472,6 @@ cfg_rt! {
         /// ```
         ///
         /// [handle]: fn@Handle::block_on
-        #[track_caller]
         pub fn block_on<F: Future>(&self, future: F) -> F::Output {
             #[cfg(all(tokio_unstable, feature = "tracing"))]
             let future = crate::util::trace::task(future, "block_on", None, task::Id::next().as_u64());