+++ tokio/src/task/mod.rs
@@ -291,6 +291,11 @@ cfg_rt! {
     mod yield_now;
     pub use yield_now::yield_now;
 
+    cfg_unstable! {
+        mod consume_budget;
+        pub use consume_budget::consume_budget;
+    }
+
     mod local;
     pub use local::{spawn_local, LocalSet};
 
@@ -301,7 +306,8 @@ cfg_rt! {
     pub use unconstrained::{unconstrained, Unconstrained};
 
     cfg_unstable! {
-        mod join_set;
+        pub mod join_set;
+        #[doc(inline)]
         pub use join_set::JoinSet;
         pub use crate::runtime::task::{Id, AbortHandle};
     }